+++ R/read_delim.R
@@ -2,7 +2,7 @@
 #' @importClassesFrom Rcpp "C++Object"
 NULL
 
-#' Read a delimited file into a data frame.
+#' Read a delimited file (including csv & tsv) into a tibble
 #'
 #' \code{read_csv} and \code{read_tsv} are special cases of the general
 #' \code{read_delim}. They're useful for reading the most common types of
@@ -62,7 +62,7 @@ NULL
 #' read_csv(readr_example("mtcars.csv"))
 #' read_csv(readr_example("mtcars.csv.zip"))
 #' read_csv(readr_example("mtcars.csv.bz2"))
-#' read_csv("https://github.com/hadley/readr/raw/master/inst/extdata/mtcars.csv")
+#' read_csv("https://github.com/tidyverse/readr/raw/master/inst/extdata/mtcars.csv")
 #'
 #' # Or directly from a string (must contain a newline)
 #' read_csv("x,y\n1,2\n3,4")
@@ -208,10 +208,10 @@ generate_spec_fun <- function(x) {
   x
 }
 
-#' Retrieve the column specification of a file.
+#' Generate a column specification
 #'
-#' By default the types of the first 20 columns are printed,
-#' \code{options(readr.num_columns)} can be used to modify this (a value of 0
+#' When printed, only the first 20 columns are printed by default. To override,
+#' set \code{options(readr.num_columns)} can be used to modify this (a value of 0
 #' turns off printing).
 #'
 #' @return The \code{col_spec} generated for the file.