+++ R/source.R
@@ -23,10 +23,14 @@
 #' datasource(readr_example("mtcars.csv"))
 #' datasource(readr_example("mtcars.csv.bz2"))
 #' datasource(readr_example("mtcars.csv.zip"))
-#' datasource("https://github.com/hadley/readr/raw/master/inst/extdata/mtcars.csv")
+#' \dontrun{
+#' datasource("https://github.com/tidyverse/readr/raw/master/inst/extdata/mtcars.csv")
+#' }
 #'
 #' # Connection
-#' datasource(rawConnection(charToRaw("abc\n123")))
+#' con <- rawConnection(charToRaw("abc\n123"))
+#' datasource(con)
+#' close(con)
 datasource <- function(file, skip = 0, comment = "") {
   if (inherits(file, "source")) {
     file