+++ src/SourceRaw.h
@@ -16,6 +16,9 @@ public:
     begin_ = (const char*) RAW(x);
     end_ = (const char*) RAW(x) + Rf_xlength(x);
 
+    // Skip byte order mark, if needed
+    begin_ = skipBom(begin_, end_);
+
     // Skip lines, if needed
     begin_ = skipLines(begin_, end_, skip, comment);
   }