+++ NEWS.md
@@ -1,5 +1,11 @@
 # readr 1.0.0.9000
 
+* `col_euro_double()`, `parse_euro_double()`, `col_numeric()`, and 
+  `parse_numeric()` have been removed.
+
+* `guess_encoding()` returns a tibble, and works better with lists of 
+  raw vectors (as returned by `read_lines_raw()`).
+
 # readr 1.0.0
 
 ## Column guessing